feat: initial work done to rely on new global config dir
