+++ package.json
@@ -105,6 +105,7 @@
     "del": "^5.1.0",
     "dot-prop": "^5.1.0",
     "dotenv": "^8.2.0",
+    "env-paths": "^2.2.0",
     "envinfo": "^7.3.1",
     "execa": "^5.0.0",
     "express": "^4.17.1",
